+++ src/sagapilot/mpworker/pilot_manager_worker.py
@@ -1,0 +1,10 @@
+"""
+.. module:: sinon.pilot_manager
+   :platform: Unix
+   :synopsis: Implementation of the PilotManager class.
+
+.. moduleauthor:: Ole Weidner <ole.weidner@rutgers.edu>
+"""
+
+__copyright__ = "Copyright 2013, http://radical.rutgers.edu"
+__license__   = "MIT"