+++ src/sagapilot/tests/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# encoding: utf-8
+
+import os
+import unittest
+
+from api    import * 
+from remote import *