+++ src/sagapilot/url.py
@@ -1,0 +1,18 @@
+"""
+.. module:: sagapilot.url
+   :platform: Unix
+   :synopsis: Implementation of the Url class (forwarded from SAGA-Python).
+
+.. moduleauthor:: Ole Weidner <ole.weidner@rutgers.edu>
+"""
+
+__copyright__ = "Copyright 2013-2014, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+from saga import Url as sagaURL
+
+
+# ------------------------------------------------------------------------------
+#
+class Url(sagaURL): 
+    pass