+++ CHANGES.md
@@ -1,6 +1,20 @@
-Version 0.1.0 release 
+0.1.3 Release 
 ---------------------------------------------------------------------
 
-* Initial release (PyPi: )
+* Github only release: 
+  pip install --upgrade -e git://github.com/saga-project/saga-pilot.git@master#egg=saga-pilot 
+
+* Added logging
+* Added security credential handling 
 * Closed tickets: 
-  - https://github.com/saga-project/saga-pilot/issues?milestone=1&state=closed
+  - https://github.com/saga-project/saga-pilot/issues?milestone=3&state=closed
+
+
+0.1.2 Release 
+---------------------------------------------------------------------
+
+* Github only release: 
+  pip install --upgrade -e git://github.com/saga-project/saga-pilot.git@master#egg=saga-pilot 
+
+* Closed tickets: 
+  - https://github.com/saga-project/saga-pilot/issues?milestone=4&state=closed